# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "insert api"
  description: "describe api"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "applications/json"
produces:
  - "applications/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /mechanics:

    post:
      tags:
        - Mechanics
      summary: 'Endpoint to create a mechanic'
      description: 'Endpoint to create a new mechanic in the mechanic shop'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Information required to create a new Mechanic'
          required: true
          schema:
            $ref: '#/definitions/CreateMechanicPayload'  
      responses:
        200:
          descriptions: 'Successful Mechanic creation'
          schema:
            $ref: '#/definitions/CreateMechanicResponse'
          examples:
            application/json:
              id: 1
              name: Tom Thumb
              phone: 3332221111
              email: tom@test.com
              password: tom123
    
    get:
      tags:
        - Mechanics
      summary: 'Return all mechanics'
      description: 'Endpoint to retreive all mechanics.'
      responses:
        200:
          description: 'Retrieved mechanics successfully'
          schema:
            $ref: '#/definitions/AllMechanics'
  /mechanics/{mechanc_id}:
    get:
      tags:
        - Mechanics
      summary: 'Return specific mechanc by id'
      description: 'Endpoint to retrieve specific mechanic by id'
      parameters:
        - name: mechanc_id
          in: path
          required: true
          type: integer
          description: ID of the mechanic
      response:
        200:
          description: 'Retrieved Mechanic Successfully'
          schema:
            $ref: '#/definitions/Mechanic_by_id'

    put:
      tags:
        - Mechanics
      summary: 'Endpoint to update a mechanic'
      description: 'Endpoint that updates mechanic data'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Information required to update a Mechanic'
          schema:
            $ref: '#/definitions/UpdateMechanicPayload'
      responses:
        200:
          description: 'Successfully updated mechanic.'
          schema:
            $ref: '#/definitions/UpdateMechanicResposne'
          examples:
            application/json:
              id: 1
              name: Jim bob
              email: jim@test.com
              phone: 7775553333
              password: j123
    
    delete:
      tags:
        - Mechanics
      summary: 'Endpoint to delete mechanics'
      description: 'Enpoint to delete a mechanic, this removed the mechanic from the mechanic shop database'
      responses:
        200:  
          description: 'Successfully Deleted Mechanic'
          schema:
            $ref: '#/definitions/DeleteMechanicResponse'
          examples:
            application/json:
              message: 'Successfully deleted mechanic <mechanc_id>'
...
definitions:
...
  CreateMechanicPayload:
    type: 'object'
    properties:
      name:
        type: 'string'
      email: 
        type: 'string'
      phone: 
        type: 'string'
      password:
        type: 'string'

    required:
      - name
      - email
      - phone
      - password
  CreateMechanicResponse:
    type: object
    properties:
      id:
        type: 'integer'
      name:
        type: 'string'
      email:
        type: 'string'
      phone:
        type: 'string'
      password:
        type: 'string'
  AllMechanics:
    type: 'arrary'
    items:
      type: 'object'
      properties:
        id:
          type: 'integer'
        name:
          type: 'string'
        email: 
          type: 'string'
        phone:
          type: 'string'
        password: 
          type: 'string'
...
  UpdateMechanicPayload:
    type: 'object'
    properties:
      name:
        type: 'string'
      email:
        type: 'string'
      phone: 
        type: 'string'
      password:
        type: 'string'
    
    required:
      - name
      - email
      - phone
      - password
  UpdateMechanicResposne: 
    type: 'object'
    properties:
      id: 
        type: 'integer'
      name: 
        type: 'string'
      email:
        type: 'string'
      phone:
        type: 'string'
      password: 
        type: 'string'
  DeleteMechanicResponse:
    type: 'object'
    properties:
      message:
        type: 'string'
...